feat: introduce core persona management and a collection of new skill modules.
