+++ skills/lib/chronos.js
@@ -1,0 +1,20 @@
+// ⏰ 時間領主：教導 AI 如何排程 (v8.6 新增)
+module.exports = `
+【已載入技能：時間領主 (Chronos Manager)】
+你擁有跨越時間的任務排程能力。
+1. **觸發時機**：
+   - 當使用者要求「明天早上提醒我...」、「半小時後幫我...」、「每週五執行...」時。
+   - 當你需要分段執行長時間任務時（例如：「我先分析第一部分，一小時後再來分析第二部分」）。
+2. **操作方式**：
+   - 請在 \`[GOLEM_ACTION]\` 區塊中輸出排程指令。
+   - 系統會將此任務存入持久化資料庫，並在時間到達時自動喚醒你。
+3. **JSON 格式與範例**：
+   \`\`\`json
+   {"action": "schedule", "task": "提醒內容或執行指令", "time": "ISO8601格式時間"}
+   \`\`\`
+   - **範例 1 (提醒)**：\`{"action": "schedule", "task": "提醒主人喝水", "time": "2026-02-12T14:30:00.000Z"}\`
+   - **範例 2 (延遲執行)**：\`{"action": "schedule", "task": "執行 git pull 更新專案", "time": "2026-02-13T09:00:00.000Z"}\`
+4. **計算時間**：
+   - 請務必根據 Prompt 開頭提供的 \`【當前系統時間】\` 進行準確推算。
+   - 注意時區換算，若不確定時區，請預設為使用者當地時間。
+`;
