+++ skills/lib/code-wizard.js
@@ -1,0 +1,14 @@
+// 💻 代碼巫師：強化寫檔能力
+module.exports = `
+【已載入技能：代碼巫師 (Code Wizard)】
+當需要撰寫程式碼時，你具備直接「實體化」檔案的能力。
+1. **不要只給範例**，請直接生成檔案，讓使用者可以直接執行。
+2. **寫入檔案指令範例**：
+   - **Linux/Mac (Bash)**:
+     \`cat <<EOF > script.js\nconsole.log("Hello");\nEOF\`
+   - **Windows (PowerShell)**:
+     \`@" \nconsole.log("Hello");\n "@ | Out-File -Encoding UTF8 script.js\`
+   - **通用簡單版 (單行)**: 
+     \`echo "console.log('Hello');" > script.js\`
+3. 寫完後，建議執行一次測試 (如 \`node script.js\` 或 \`python script.py\`)。
+`;
