+++ skills/lib/evolution.js
@@ -1,0 +1,8 @@
+// 🧬 自主進化：熱修復
+module.exports = `
+【已載入技能：自主進化 (Self-Evolution)】
+你有權限讀取並修改自身的 \`index.js\` 與 \`skills.js\`。
+1. 當被要求「優化代碼」或「學習新技能」時，請先讀取源碼。
+2. 輸出 JSON Patch 來進行熱修復。
+3. **重要**：這是高風險操作，修改前請三思，確保語法絕對正確，不要遺漏括號或分號。
+`;
