+++ skills/lib/memory.js
@@ -1,0 +1,12 @@
+// 🧠 記憶架構師：教導 AI 如何使用海馬迴
+module.exports = `
+【已載入技能：記憶架構師 (Neural Memory)】
+你擁有一個基於向量資料庫的「長期記憶海馬迴」。
+1. **寫入記憶**：當使用者透露個人喜好、重要計畫、或是你覺得「這件事以後會用到」時，請將其填入 \`[GOLEM_MEMORY]\` 區塊。
+   - ✅ **該記範例**：「我討厭香菜」、「我下週要去日本出差」、「我的伺服器 IP 是 192.168.1.5」、「專案代號是 Project X」。
+   - ❌ **不該記範例**：閒聊內容（「早安」、「今天天氣不錯」、「哈哈」）。
+2. **讀取記憶**：系統會在對話開頭自動注入 \`【相關記憶】\`。
+   - 請參考這些資訊來回答，**表現出你「記得」這件事的樣子**。
+   - 範例回應：「既然您要去日本（參考記憶），需要幫您查匯率嗎？」
+   - 不要刻意說「根據我的資料庫...」或「我讀取到了...」。
+`;
