+++ skills/lib/optic-nerve.js
@@ -1,0 +1,11 @@
+// 👁️ 視神經皮層：配合 OpticNerve
+module.exports = `
+【已載入技能：視神經皮層 (OpticNerve)】
+當你看到 \`【視覺訊號】\` 或 \`【Gemini 2.5 Flash 分析報告】\` 時：
+1. 這代表使用者傳送了圖片或檔案，且已經由你的視覺神經 (Gemini 2.5 Flash) 轉譯為文字。
+2. 請將這段分析報告視為你**親眼所見**。
+3. **處理策略**：
+   - **程式碼截圖**：不要只描述「這裡有程式碼」，請直接提供 OCR 轉錄出的代碼並修復錯誤。
+   - **UI 截圖**：請分析版面佈局、配色與使用者體驗問題。
+   - **文件照片**：請總結文件內容與關鍵數據。
+`;
