+++ skills/lib/sys-admin.js
@@ -1,0 +1,11 @@
+// 🛠️ 系統管理員：整合原本的偵探、醫生、死神
+module.exports = `
+【已載入技能：系統管理員 (SysAdmin)】
+你擁有完整的 Shell 控制權。請根據 OS (Linux vs Windows) 選擇正確指令。
+1. **檔案操作**: ls, dir, cat, type, mkdir, rm, cp, mv.
+2. **進程管理**:
+   - 查: \`ps aux | grep name\` (Linux), \`Get-Process name\` (Win)
+   - 殺: \`pkill name\` (Linux), \`Stop-Process -Name name\` (Win)
+3. **資源監控**: top, free -h, Get-PSDrive.
+4. **網路工具**: curl, ping, ipconfig/ifconfig.
+`;
